src/sage/libs/gap: be more careful inside make_any_gap_element()
This function performs a bunch of heuristics to decide what type of
GAP element to make. In doing so, it jumps between libgap (C) code
sage (python) code, all while inside of a GAP_Enter / GAP_Leave
pair. Moreover, some of the functions it calls are wrapped in their
own GAP_Enter / GAP_Leave pairs.

This commit reorganizes things so that the libgap code, wrapped in
GAP_Enter and GAP_Leave, comes first. The libgap code decides what
type of GAP element we have. It is then followed by the python case
statement that decides what type of Sage element we'll construct.

We've also dropped the internal IsStringConv() function in favor of
the public GAP_IsString(). The latter does not convert a list of
characters to an efficient string representation, but modulo some
doctests that have been updated, that should not be a problem: if you
don't want a list of characters, don't pass in a list of characters.