src/sage/libs/gap/element.pyx: more tolerance in Ctrl-C tests
The check for _pow_ being interruptible by Ctrl-C is subject to some
variation due to CPU usage, scheduling, etc. Here we increase the
tolerance to 5s, which should be more than enough time to interrupt
it, and also increase the size of the problem so that it runs for
at least 5s even on fast machines.