src/sage/rings/complex_arb.pyx: add some Ctrl-C tolerance to a doctest
We have one doctest in this file that is perpetually failing on the CI
because a CBF integral takes too long to interrupt. The computation is
extremely long-running, though, and "too long" is on the order of
deciseconds: everything is fine. This commit increases the wait
tolerance from 0.2s to 5s. That should be long enough to avoid all
random CPU scheduling issues, and is still much shorter than the time
it would take for the integral to complete.