src/sage/doctest/sources.py: add "long time" for two tests
Two calls to FileDocTestSource.create_doctests() in this file can take
a while (about 5s) on the first run. This now triggers a "slow
doctest" warning if you are unlucky, and is unusually annoying because
doctesting this file is expected to produce no output in a doctest in
src/sage/doctest/forker.py.

We add "# long time" to correct that random doctest failure. In any
case, the two tests do take a long time.