src/sage/symbolic/integration/integral.py: delete a few giac doctests
Several tests in this file are computing integrals that "can only be
solved by giac." But since re-enabling abs_integrate for maxima, that
is no longer true: maxima is able to solve them, and the answer is
often an improvement (defined on a larger set, for example).

Since these examples were meant to demonstrate the fallback to giac,
we delete them rather than update the expected output to conform with
maxima's. We have enough trouble updating the existing maxima
expressions every time a new version of maxima is released.