src/sage/interfaces/maxima_abstract.py: reimplement _commands()
The maxima _commands() method was implemented in terms of the
tab-completion method, completions(), which goes one-letter-at-a-time
and is very slow as a result. We reimplement it directly to call
maxima's apropos() with the empty string. This gets all commands in
one shot. We also clean up the list of commands a little by excluding
some junk results. The diff (eliminated elements) from the old list to
the new one is,

  {'SPLITS\\IN\\Q',
   'erf_%iargs',
   'exp-form',
   'is-boole-eval',
   'is-boole-verify',
   'maybe-boole-eval',
   'maybe-boole-verify',
   'sstatus-aux',
   'time\\/\\/call',
   'unknown\\?',
   'zeta%pi'}

The faster method should help avoid warnings in the CI about the test
being too slow.