src/sage/matrix/matrix_integer_dense_hnf.py: more reliable HNF
We have a random test for the HNF of an integer dense matrix that is
failing on one particular example. The HNF algorithm is probabilistic,
and we call probable_pivot_columns() expecting the matrix consisting
of those columns to be square. But if you are unlucky, the reduction
modulo a random prime in probable_pivot_columns() can reduce the rank
of the matrix, and you'll get something non-square. This kills the
hnf_square() function that is called shortly thereafter.

For the fix, we simply loop until probable_pivot_columns() does in
fact give us a square matrix. Since at each iteration a random prime
in the range (10007, 46000) is used, it should (ha ha) succeed
eventually.