src/sage/modular/modform/numerical.py: increase tolerance, again
To avoid randomly failing tests, we bump the tolerance on some tests
for numerical_eigenforms(). These tolerances have been increased in
the past without any formal motivation, and we are still on the order
of 1e-11 (Which Should Be Good Enough For Anyone), so to fix the tests
we simply increase the tolerance again.