src/sage/matrix/matrix_gf2e_dense.pyx: speed up inverse example
We have an example in this file demonstrating is_invertible() for an
element of GL(2^8, GF(2^8)). This is raising "slow doctest" warnings
on the CI and on my own machine, but the exponent is not critical to
the example. Lowering the "8" to a "6" makes the test complete much
more quickly.