src/sage/stats/distributions/discrete_gaussian_lattice.py: add long time
There's a loop in one of the examples in this file that samples from a
distribution until a condition is met. This already takes a long time
on my machine, but in the absence of luck it can take such a long time
that it triggers a "slow doctest" warning on the CI. Since the time
taken is random, we mark it "# long time" to be safe.