src/sage/stats/distributions/discrete_gaussian_lattice.py: more long time
Add "# long time" to two _normalisation_factor_zz() examples. These
take ~30s on one of my machines, and ~50s on another. Although they
run fast enough on the CI, I think it is reasonable to mark them as
taking a long time -- they are the two slowest examples in this file
for me.