sage.rings.valuation.mapped_valuation_test: add pytests
Copy two TestSuite() runs from a TESTS:: block to pytest. These are on
the verge of hitting the "slow doctest" limit for long tests, and
sometimes cause warnings in the CI. But more importantly, they are
simply better suited to pytest: the user would never want to run them,
and the output is not meaningful as documentation.