sage.graphs.graph: make is_projective_planar() a @cached_method
The is_projective_planar() function can be very slow. The first
example in the documentation is the Petersen graph, on which the
method is taking over six minutes to arrive at the answer (True).
We mark it as a @cached_method, since that will allow us to pre-
fill the cache for graphs that are known to be projective planar,
saving the user a lot of time.