sage.geometry.hyperplane_arrangement.ordered_arrangement: add "long time"
One projective_meridians() test in this file is just barely triggering
a "slow doctest" warning on the CI, at 5.04s cpu. On my (slow) machine
it takes a bit longer, around 6.5s cpu. This is longer than the
recommended length for a doctest in any case, so we mark it "long
time" to avoid warnings.