sage.rings.polynomial.multi_polynomial_libsingular: add "long time"
There's a series of tests in this file that set an alarm interrupt,
and then attempt to express a polynomial in terms of the generators of
an ideal (to which it belongs). This is taking around 5s on the CI,
and as a result, is sometimes raising "slow doctest" warnings. On my
(slow) machine, the lift takes about 15s, and the added 0.5s alarm
increases that. We mark all of these tests "long time" to be on the
safe side.