src/sage/structure/coerce.pyx: support both 1.3 and 1.4
To support both versions of mpmath, we create an internal copy of the
existing py_scalar_parent() function, and update the original to
support mpmath-1.4.x. If mpmath.__version__ indicates that v1.3 is in
use, we punt to the old copy of the function.

The copy/paste is ugly, but considering that it's only temporary,
emphasis is placed on how easy it will be to undo eventually.