src/sage/geometry/cone_catalog.py: add barker_foran()
Though it is a name I just made up, the newly minted Barker-Foran cone
is a polyhedral self-dual convex cone with five extreme rays in three
dimensions. Barker & Foran give it as an example in their paper, which
we cite.

It is a nice example because asymmetric self-dual cones are hard to
come by.